Update Monetization & Licensing section in index.md
Rewrote the Monetization & Licensing section in the index.md file to offer a clearer perspective on our current stance and philosophy regarding the mentioned aspects. The previous paragraph was brief and leaned towards informing about a free-to-use model while exploring monetization for sustainability and innovation.

The new write-up vividly elaborates on 'Intrig's approach to accessibility, innovation, monetization and the importance of licensing. It seeks to establish as an open, collaborative, ethical project with a community-driven approach while amicably introducing a sense of anticipation about impending updates in the licensing model.